+++ packages/plugins/embeds/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @emdash-cms/plugin-embeds
 
+## 0.1.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.26.0
+
+## 0.1.31
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.25.1
+
+## 0.1.30
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.25.0
+
+## 0.1.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.24.1
+
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.24.0
+
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.23.0
+
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.22.0
+
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.21.0
+
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.20.0
+
+## 0.1.23
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.19.0
+
+## 0.1.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.18.0
+
 ## 0.1.21
 
 ### Patch Changes